Consolidate utility methods
