+++ app/src/main/java/org/ergoplatform/android/ErgoFacade.kt
@@ -1,7 +1,6 @@
 package org.ergoplatform.android
 
 import StageConstants
-import android.content.Context
 import android.util.Log
 import com.google.gson.Gson
 import com.google.gson.JsonObject
@@ -10,11 +9,7 @@ import org.ergoplatform.appkit.*
 import org.ergoplatform.appkit.impl.ErgoTreeContract
 import org.ergoplatform.wallet.mnemonic.WordList
 import scala.collection.JavaConversions
-import java.math.RoundingMode
-import java.text.DecimalFormat
-import java.text.DecimalFormatSymbols
 import java.util.*
-import kotlin.math.ln
 import kotlin.math.pow
 
 val MNEMONIC_WORDS_COUNT = 15
@@ -32,54 +27,11 @@ fun ergsToNanoErgs(ergs: Double): Long {
     return nanoergs
 }
 
-/**
- * ERG is always formatted US-style (e.g. 1,000.00)
- */
-fun formatErgsToString(ergs: Double, context: Context): String {
-    return DecimalFormat(context.getString(R.string.format_erg), DecimalFormatSymbols(Locale.US)).format(ergs)
-}
-
-/**
- * fiat is formatted according to users locale, because it is his local currency
- */
-fun formatFiatToString(amount: Double, currency: String, context: Context): String {
-    return DecimalFormat(context.getString(R.string.format_fiat)).format(amount) +
-            " " + currency.toUpperCase(Locale.getDefault())
-}
-
-/**
- * Formats token (asset) amounts, always formatted US-style
- *
- * @param formatWithPrettyReduction 1,120.00 becomes 1.1K, useful for displaying with less space
- */
-fun formatTokenAmounts(
-    amount: Long,
-    decimals: Int,
-    formatWithPrettyReduction: Boolean = false
-): String {
-    val valueToShow: Double = longWithDecimalsToDouble(amount, decimals)
-
-    return if (valueToShow < 1000 || !formatWithPrettyReduction) {
-        ("%." + (Math.min(5, decimals)).toString() + "f").format(Locale.US, valueToShow)
-    } else {
-        formatDoubleWithPrettyReduction(valueToShow)
-    }
-}
-
-fun formatDoubleWithPrettyReduction(amount: Double): String {
-    val suffixChars = "KMGTPE"
-    val formatter = DecimalFormat("###.#", DecimalFormatSymbols(Locale.US))
-    formatter.roundingMode = RoundingMode.DOWN
-
-    return if (amount < 1000.0) formatter.format(amount)
-    else {
-        val exp = (ln(amount) / ln(1000.0)).toInt()
-        formatter.format(amount / 1000.0.pow(exp.toDouble())) + suffixChars[exp - 1]
-    }
-}
-
 fun longWithDecimalsToDouble(amount: Long, decimals: Int) =
     (amount.toDouble()) / (10.0.pow(decimals))
+
+fun doubleToLongWithDecimals(amount: Double, decimals: Int) =
+    (amount * 10.0.pow(decimals)).toLong()
 
 fun serializeSecrets(mnemonic: String): String {
     val gson = Gson()
