+++ app/src/main/java/org/ergoplatform/android/ui/UiUtils.kt
@@ -11,6 +11,14 @@ import androidx.navigation.NavController
 import androidx.navigation.NavDirections
 import androidx.navigation.NavOptions
 import androidx.navigation.Navigator
+import org.ergoplatform.android.R
+import org.ergoplatform.android.longWithDecimalsToDouble
+import java.math.RoundingMode
+import java.text.DecimalFormat
+import java.text.DecimalFormatSymbols
+import java.util.*
+import kotlin.math.ln
+import kotlin.math.pow
 
 fun forceShowSoftKeyboard(context: Context) {
     ContextCompat.getSystemService(
@@ -31,6 +39,52 @@ fun inputTextToDouble(amountStr: String?): Double {
         return if (amountStr.isNullOrEmpty()) 0.0 else amountStr.toDouble()
     } catch (t: Throwable) {
         return 0.0
+    }
+}
+
+/**
+ * ERG is always formatted US-style (e.g. 1,000.00)
+ */
+fun formatErgsToString(ergs: Double, context: Context): String {
+    return DecimalFormat(context.getString(R.string.format_erg), DecimalFormatSymbols(Locale.US)).format(ergs)
+}
+
+/**
+ * fiat is formatted according to users locale, because it is his local currency
+ */
+fun formatFiatToString(amount: Double, currency: String, context: Context): String {
+    return DecimalFormat(context.getString(R.string.format_fiat)).format(amount) +
+            " " + currency.toUpperCase(Locale.getDefault())
+}
+
+/**
+ * Formats token (asset) amounts, always formatted US-style
+ *
+ * @param formatWithPrettyReduction 1,120.00 becomes 1.1K, useful for displaying with less space
+ */
+fun formatTokenAmounts(
+    amount: Long,
+    decimals: Int,
+    formatWithPrettyReduction: Boolean = false
+): String {
+    val valueToShow: Double = longWithDecimalsToDouble(amount, decimals)
+
+    return if (valueToShow < 1000 || !formatWithPrettyReduction) {
+        ("%." + (Math.min(5, decimals)).toString() + "f").format(Locale.US, valueToShow)
+    } else {
+        formatDoubleWithPrettyReduction(valueToShow)
+    }
+}
+
+fun formatDoubleWithPrettyReduction(amount: Double): String {
+    val suffixChars = "KMGTPE"
+    val formatter = DecimalFormat("###.#", DecimalFormatSymbols(Locale.US))
+    formatter.roundingMode = RoundingMode.DOWN
+
+    return if (amount < 1000.0) formatter.format(amount)
+    else {
+        val exp = (ln(amount) / ln(1000.0)).toInt()
+        formatter.format(amount / 1000.0.pow(exp.toDouble())) + suffixChars[exp - 1]
     }
 }
